Add ability to shut specific workspace
In this change we introduce the ability to shut specific workspaces
with the command "shut <name>". In order to shut all workspaces
we now need to use "shut all". I tried to make just "shut" work
but wasn't obvious.

This works by just adding the workspace as an attribute to the
minion objects, which we can then easily filter by when closing windows.